commit the stash

+++ HealthCar3/ConsoleApp1/data/components/PanelComponent.cs
@@ -6,14 +6,14 @@ namespace ConsoleApp1.data.components
 {
     class PanelComponent
     {
-        public int[] size;
+        public double[] size;
         public int[] resolution;
         public int[] background;
         public bool castShadow;
 
-        public PanelComponent(int height, int width, int resx, int resy, int bg1, int bg2, int bg3, int bg4, bool castShadow)
+        public PanelComponent(double height, double width, int resx, int resy, int bg1, int bg2, int bg3, int bg4, bool castShadow)
         {
-            this.size = new int[] { height, width };
+            this.size = new double[] { height, width };
             this.resolution = new int[] { resx, resy };
             this.background = new int[] { bg1, bg2, bg3, bg4 };
             this.castShadow = castShadow;
